Split the common classes as well
