+++ importer/Importers/SmfCommonOriginStep3.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * @name      OpenImporter
+ * @copyright OpenImporter contributors
+ * @license   BSD http://opensource.org/licenses/BSD-3-Clause
+ *
+ * @version 2.0 Alpha
+ *
+ * This file contains code based on:
+ *
+ * Simple Machines Forum (SMF)
+ * copyright:	2011 Simple Machines (http://www.simplemachines.org)
+ * license:	BSD, See included LICENSE.TXT for terms and conditions.
+ */
+
+namespace OpenImporter\Importers;
+
+abstract class SmfCommonOriginStep3 extends Step3BaseImporter
+{
+	public function run($import_script)
+	{
+		$to_prefix = $this->config->to_prefix;
+
+		// add some importer information.
+		$this->db->query("
+			REPLACE INTO {$to_prefix}settings (variable, value)
+				VALUES ('import_time', " . time() . "),
+					('enable_password_conversion', '1'),
+					('imported_from', '" . $import_script . "')");
+	}
+}
